🛠️ Refactor: User, jwt HTTP 요청/응답 처리 방식 리펙토링
1. Dto -> request / response 방식으로 변경

Related to: #29
