+++ src/main/java/darkoverload/itzip/user/controller/request/EmailSendRequest.java
@@ -1,4 +1,4 @@
-package darkoverload.itzip.user.dto;
+package darkoverload.itzip.user.controller.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
@@ -10,7 +10,7 @@ import lombok.Getter;
  */
 @Getter
 @AllArgsConstructor
-public class EmailSendDto {
+public class EmailSendRequest {
     @NotEmpty(message = "이메일을 입력해주세요.")
     @Pattern(regexp = "^[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*\\.[a-zA-Z]{2,3}$",
             message = "올바르지 않은 이메일 형식입니다.")
