+++ src/main/java/darkoverload/itzip/user/controller/request/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package darkoverload.itzip.user.dto;
+package darkoverload.itzip.user.controller.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
@@ -8,7 +8,7 @@ import lombok.*;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserLoginDto {
+public class UserLoginRequest {
     @NotEmpty(message = "이메일을 입력해주세요.")
     @Pattern(regexp = "^[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\\.]?[0-9a-zA-Z])*\\.[a-zA-Z]{2,3}$",
             message = "올바르지 않은 이메일 형식입니다.")
